# src/main/resources/static/swagger-ui/openapi.yaml
openapi: 3.0.0
info:
  title: API de Residencias
  version: 1.0.0
  description: API para gestionar residencias, con operaciones de registro y consulta.
paths:
...
  /resi/{idResidencia}/juego/{idJuego}/get:
    get:
      tags:
        - Modulo - JUEGO
      summary: Obtener juego
      description: Devuelve un juego.
      parameters:
        - name: idResidencia
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: idJuego
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Juego(s) encontrado(s)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Juego.ResponseDto'
        '400':
          description: Datos nulos/vacios de la variable de la url
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                CAMPOS_OBLIGATORIOS:
                  value:
                    mensaje: Faltan campos obligatorios
                    codigo: 1001
                    status: 400
                    timestamp: 2025-04-13T21:00:00
        '404':
          description: Juego invalida, no existe en la base de datos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                JUEGO_INVALIDO:
                  value:
                    mensaje: Juego invalida
                    codigo: 1008
                    status: 404
                    timestamp: 2025-04-13T21:00:00
        '405':
          description: Este endPoint esta protegido por ciertas cosas, debe tener un token. Que se le proporciona cada vez que hace  login.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                ENDPOINT_PROTEGIDO:
                  value:
                    mensaje: Tiene que tener permiso para acceder aqui.
                    codigo: 1020
                    status: 405
                    timestamp: 2025-04-13T21:00:0
  /resi/{idResidencia}/juego/getAll:
    get:
      tags:
        - Modulo - JUEGO
      summary: Obtener lista de juego
      description: Devuelve una lista de juegos.
      parameters:
        - name: idResidencia
          in: query
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JuegoFiltre.Dto'
      responses:
        '200':
          description: Juego(s) encontrado(s)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Juego.ResponseDto'
        '400':
          description: Datos nulos/vacios de la variable de la url
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                CAMPOS_OBLIGATORIOS:
                  value:
                    mensaje: Faltan campos obligatorios
                    codigo: 1001
                    status: 400
                    timestamp: 2025-04-13T21:00:00
        '404':
          description: Residencia invalida, no existe en la base de datos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                RESIDENCIA_INVALIDO:
                  value:
                    mensaje: Residencia invalida
                    codigo: 1006
                    status: 404
                    timestamp: 2025-04-13T21:00:00
        '405':
          description: Este endPoint esta protegido por ciertas cosas, debe tener un token. Que se le proporciona cada vez que hace  login.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                ENDPOINT_PROTEGIDO:
                  value:
                    mensaje: Tiene que tener permiso para acceder aqui.
                    codigo: 1020
                    status: 405
                    timestamp: 2025-04-13T21:00:0
  /resi/{idResidencia}/juego/{idJuego}/delete:
    delete:
      tags:
        - Modulo - JUEGO
      summary: Eliminar una Juego.
      description: Elimina un juego en cascada, con todas las entidades relacionadas a el.
      parameters:
        - name: idResidencia
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: idJuego
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Juego borrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Juego.ResponseDto'
        '400':
          description: Datos nulos/vacios de la variable de la url
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                CAMPOS_OBLIGATORIOS:
                  value:
                    mensaje: Faltan campos obligatorios
                    codigo: 1001
                    status: 400
                    timestamp: 2025-04-13T21:00:00
        '404':
          description: Juego invalido, no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                RESIDENCIA_INVALIDO:
                  value:
                    mensaje: Juego invalido
                    codigo: 1008
                    status: 404
                    timestamp: 2025-04-13T21:00:00
        '405':
          description: Este endPoint esta protegido por ciertas cosas, debe tener un token. Que se le proporciona cada vez que hace  login.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                ENDPOINT_PROTEGIDO:
                  value:
                    mensaje: Tiene que tener permiso para acceder aqui.
                    codigo: 1020
                    status: 405
                    timestamp: 2025-04-13T21:00:00
...
  /resi/{idResidencia}/juego/{idJuego}/registro/{idRegistroJuego}/get:
    get:
      tags:
        - Modulo - JUEGO
      summary: Obtener registro de juego
      description: Devuelve los datos de una partida específica.
      security:
        - bearerAuth: []
      parameters:
        - name: idResidencia
          in: path
          required: true
          schema:
            type: integer
        - name: idJuego
          in: path
          required: true
          schema:
            type: integer
        - name: idRegistroJuego
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Registro encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistroJuego.ResponseDto'
        '400':
          description: Datos nulos/vacios .
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                CAMPOS_OBLIGATORIOS:
                  value:
                    mensaje: Faltan campos obligatorios
                    codigo: 1001
                    status: 400
                    timestamp: 2025-04-13T21:00:00
        '404':
          description: juego/ registro_juego invalido, no existe en la base de datos.
          content:
            application/json:
              schema:
                  $ref: '#/components/schemas/Error.ResponseDto'
              examples:
                JUEGO_INVALIDO:
                  value:
                    mensaje: Juego invalida
                    codigo: 1008
                    status: 404
                    timestamp: 2025-04-13T21:00:00
                REGISTRO_JUEGO_INVALIDO:
                  value:
                    mensaje: Registro Juego invalido
                    codigo: 1012
                    status: 404
                    timestamp: 2025-04-13T21:00:00
        '405':
          description: Este endPoint esta protegido por ciertas cosas, debe tener un token. Que se le proporciona cada vez que hace  login.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                ENDPOINT_PROTEGIDO:
                  value:
                    mensaje: Tiene que tener permiso para acceder aqui.
                    codigo: 1020
                    status: 405
                    timestamp: 2025-04-13T21:00:00
...
  /resi/{idResidencia}/juego/{idJuego}/registro/{idRegistroJuego}/addComment:
    patch:
      tags:
        - Modulo - JUEGO
      summary: Añadir comentario a registro
      description: Actualiza la observación/comentario del registro.
      security:
        - bearerAuth: []
      parameters:
        - name: idResidencia
          in: path
          required: true
          schema:
            type: integer
        - name: idJuego
          in: path
          required: true
          schema:
            type: integer
        - name: idRegistroJuego
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistroJuego.UpdateDto'
      responses:
        '200':
          description: Comentario actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegistroJuego.ResponseDto'
        '400':
          description: Datos nulos/vacios .
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                CAMPOS_OBLIGATORIOS:
                  value:
                    mensaje: Faltan campos obligatorios
                    codigo: 1001
                    status: 400
                    timestamp: 2025-04-13T21:00:00
        '404':
          description: juego/ registro_juego invalido, no existe en la base de datos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              examples:
                JUEGO_INVALIDO:
                  value:
                    mensaje: Juego invalida
                    codigo: 1008
                    status: 404
                    timestamp: 2025-04-13T21:00:00
                REGISTRO_JUEGO_INVALIDO:
                  value:
                    mensaje: Registro Juego invalido
                    codigo: 1012
                    status: 404
                    timestamp: 2025-04-13T21:00:00
        '405':
          description: Este endPoint esta protegido por ciertas cosas, debe tener un token. Que se le proporciona cada vez que hace  login.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                ENDPOINT_PROTEGIDO:
                  value:
                    mensaje: Tiene que tener permiso para acceder aqui.
                    codigo: 1020
                    status: 405
                    timestamp: 2025-04-13T21:00:00
  /resi/{idResidencia}/juego/{idJuego}/registro/{idRegistroJuego}/delete:
    delete:
      tags:
        - Modulo - JUEGO
      summary: Eliminar registro de juego
      description: Elimina un registro de juego de forma permanente.
      security:
        - bearerAuth: []
      parameters:
        - name: idResidencia
          in: path
          required: true
          schema:
            type: integer
        - name: idJuego
          in: path
          required: true
          schema:
            type: integer
        - name: idRegistroJuego
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Registro eliminado correctamente
        '400':
          description: Datos nulos/vacios .
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                CAMPOS_OBLIGATORIOS:
                  value:
                    mensaje: Faltan campos obligatorios
                    codigo: 1001
                    status: 400
                    timestamp: 2025-04-13T21:00:00
        '404':
          description: juego/ registro_juego invalido, no existe en la base de datos.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              examples:
                JUEGO_INVALIDO:
                  value:
                    mensaje: Juego invalida
                    codigo: 1008
                    status: 404
                    timestamp: 2025-04-13T21:00:00
                REGISTRO_JUEGO_INVALIDO:
                  value:
                    mensaje: Registro Juego invalido
                    codigo: 1012
                    status: 404
                    timestamp: 2025-04-13T21:00:00
        '405':
          description: Este endPoint esta protegido por ciertas cosas, debe tener un token. Que se le proporciona cada vez que hace  login.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error.ResponseDto'
              example:
                ENDPOINT_PROTEGIDO:
                  value:
                    mensaje: Tiene que tener permiso para acceder aqui.
                    codigo: 1020
                    status: 405
                    timestamp: 2025-04-13T21:00:00
...
components:
  schemas:
...
    JuegoFiltre.Dto:
      type: object
      properties:
        nombre:
          type: string
          example: tres en ralla
        maxRegistro:
          type: boolean
          example: true
    Juego.ResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        nombre:
          type: string
          example: Memoria Visual
        idResidencia:
          type: integer
          format: int64
          example: 3
...
    RegistroJuego.UpdateDto:
      type: object
      properties:
        observacion:
          type: string
          example: Muy lento
    RegistroJuego.ResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 42
        idResidente:
          type: integer
          format: int64
          example: 5
        idJuego:
          type: integer
          format: int64
          example: 1
        idUsario:
          type: integer
          format: int64
          example: 3
        num:
          type: integer
          example: 2
        duracion:
          type: number
          format: double
          example: 120.5
        dificultad:
          type: integer
          format: int64
          example: 1
        fecha:
          type: string
          format: date-time
          example: "2025-04-16T10:15:30"
...
    Error.ResponseDto:
      type: object
      properties:
        mensaje:
          type: string
        codigo:
          type: integer
        status:
          type: integer
        timestamp:
          type: string
          format: date-time
      example:
        mensaje: Faltan campos obligatorios
        codigo: 1001
        status: 400
        timestamp: 2025-04-13T21:00:00
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT